Int quantity checked

+++ src/shared/infra/typeorm/migrations/1612140100980-CreateOrdersProducts.ts
@@ -22,7 +22,7 @@ export default class CreateOrdersProducts1612140100980
           },
           {
             name: 'quantity',
-            type: 'decimal',
+            type: 'int',
           },
           {
             name: 'created_at',
